chore: move Constants/Constants.hpp to Utilities/Constants.hpp
