+++ src/CMakeLists.txt
@@ -12,7 +12,7 @@ target_include_directories(kilo
 
 target_sources(kilo
     PRIVATE
-        "${PROJECT_SOURCE_DIR}/src/Constants/Constants.hpp"
+        "${PROJECT_SOURCE_DIR}/src/Utilities/Constants.hpp"
         "${PROJECT_SOURCE_DIR}/src/Editor/Cursor/Cursor.hpp"
 
         "${PROJECT_SOURCE_DIR}/src/Editor/Editor.hpp"
