+++ src/IO/IO.cpp
@@ -23,7 +23,7 @@
 
 #include "IO.hpp"
 
-#include "Constants/Constants.hpp"
+#include "Utilities/Constants.hpp"
 #include <system_error>
 
 #include <array>
